Clean import utilities
