+++ optimum/exporters/onnx/utils.py
@@ -20,9 +20,8 @@ from typing import TYPE_CHECKING, Dict, Tuple, Union
 import packaging
 from transformers.utils import is_tf_available, is_torch_available
 
-from ...utils import is_diffusers_available
+from ...utils import is_diffusers_available, ORT_QUANTIZE_MINIMUM_VERSION, TORCH_MINIMUM_VERSION
 from ..tasks import TasksManager
-from .import_utils import MIN_TORCH_VERSION, ORT_QUANTIZE_MINIMUM_VERSION
 
 
 if TYPE_CHECKING:
