+++ optimum/utils/__init__.py
@@ -11,78 +11,20 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import importlib.util
-import inspect
-from contextlib import contextmanager
-
-from packaging import version
 
 
-CONFIG_NAME = "config.json"
-
-_onnxruntime_available = importlib.util.find_spec("onnxruntime") is not None
-_pydantic_available = importlib.util.find_spec("pydantic") is not None
-_accelerate_available = importlib.util.find_spec("accelerate") is not None
-_diffusers_available = importlib.util.find_spec("diffusers") is not None
-
-
-def is_onnxruntime_available():
-    try:
-        # Try to import the source file of onnxruntime - if you run the tests from `tests` the function gets
-        # confused since there a folder named `onnxruntime` in `tests`. Therefore, `_onnxruntime_available`
-        # will be set to `True` even if not installed.
-        mod = importlib.import_module("onnxruntime")
-        inspect.getsourcefile(mod)
-    except:
-        return False
-    return _onnxruntime_available
-
-
-def is_pydantic_available():
-    return _pydantic_available
-
-
-def is_accelerate_available():
-    return _accelerate_available
-
-
-def is_diffusers_available():
-    return _diffusers_available
-
-
-@contextmanager
-def check_if_pytorch_greater(target_version: str, message: str):
-    r"""
-    A context manager that does nothing except checking if the PyTorch version is greater than `pt_version`
-    """
-    import torch
-
-    if not version.parse(torch.__version__) >= version.parse(target_version):
-        raise ImportError(
-            f"Found an incompatible version of PyTorch. Found version {torch.__version__}, but only {target_version} and above are supported. {message}"
-        )
-    try:
-        yield
-    finally:
-        pass
-
-
-def check_if_transformers_greater(target_version: str) -> bool:
-    """
-    Checks whether the current install of transformers is greater than or equal to the target version.
-
-    Args:
-        target_version (str): version used as the reference for comparison.
-
-    Returns:
-        bool: whether the check is True or not.
-    """
-    import transformers
-
-    return version.parse(transformers.__version__) >= version.parse(target_version)
-
-
-from .input_generators import (  # noqa
+from .import_utils import (
+    ORT_QUANTIZE_MINIMUM_VERSION,
+    TORCH_MINIMUM_VERSION,
+    check_if_pytorch_greater,
+    check_if_transformers_greater,
+    is_accelerate_available,
+    is_diffusers_available,
+    is_onnxruntime_available,
+    is_pydantic_available,
+    torch_version,
+)
+from .input_generators import (
     DEFAULT_DUMMY_SHAPES,
     DummyAudioInputGenerator,
     DummyBboxInputGenerator,
@@ -95,7 +37,7 @@ from .input_generators import (  # noqa
     DummyTextInputGenerator,
     DummyVisionInputGenerator,
 )
-from .normalized_config import (  # noqa
+from .normalized_config import (
     NormalizedConfig,
     NormalizedConfigManager,
     NormalizedSeq2SeqConfig,
@@ -103,3 +45,6 @@ from .normalized_config import (  # noqa
     NormalizedTextConfig,
     NormalizedVisionConfig,
 )
+
+
+CONFIG_NAME = "config.json"
